+++ nntrainer/compiler/meson.build
@@ -14,7 +14,10 @@ if get_option('enable-tflite-interpreter')
                                command: [flatc, '-c', '@INPUT@'])
 
   nntrainer_sources += flat_header
-  compiler_sources += 'tflite_interpreter.cpp'
+  compiler_sources += [
+    'tflite_interpreter.cpp',
+    'tflite_opnode.cpp'
+  ]
 endif
 
 foreach s : compiler_sources
